Rename: logging -> writer Modify: Assert when states and actions are not batched Add __init__.py to tests dirs

+++ rlil/nn/tests/nn_test.py
@@ -64,7 +64,7 @@ class TestNN(unittest.TestCase):
         )
 
     def test_tanh_action_bound(self):
-        space=gym.spaces.Box(np.array([-1.0, 10.0]), np.array([1, 20]))
+        space=gym.spaces.Box(np.array([-1.0, 10.0]), np.array([1, 20]),dtype=np.float32)
         model=nn.TanhActionBound(space)
         x=torch.tensor([[100.0, 100], [-100, -100], [-100, 100], [0, 0]])
         tt.assert_almost_equal(
